--- a/test/fx/test_partitioner_order.py
+++ b/test/fx/test_partitioner_order.py
@@ -24,17 +24,28 @@ class DummyPartitioner(CapabilityBasedPartitioner):
         )
 
 
-class SimpleModule(torch.nn.Module):
+# original graph node order is: ['x', 'add', 'add_1', 'output']
+class AddModule(torch.nn.Module):
     def forward(self, x):
         y = torch.add(x, x)
-        z = torch.mul(y, x)
+        z = torch.add(y, x)
         return z
 
 
 class TestPartitionerOrder(TestCase):
-    # partitoner test to check graph node order
-    def test_partitioner_order(self):
-        m = SimpleModule()
+    # partitoner test to check graph node order remains the same with the original graph after partitioning
+    def test_partitioner_graph_node_order(self):
+        m = AddModule()
+        traced_m = torch.fx.symbolic_trace(m)
+        origin_node_order = [n.name for n in traced_m.graph.nodes]
+        partions = DummyPartitioner(traced_m).propose_partitions()
+        partion_nodes = [list(partition.nodes) for partition in partions]
+        partition_node_order = [n.name for n in partion_nodes[0]]
+        self.assertTrue(partition_node_order == origin_node_order)
+
+    # partitoner test to check graph node order remains the same during multiple runs
+    def test_partitioner_multiple_runs_order(self):
+        m = AddModule()
         traced_m = torch.fx.symbolic_trace(m)
         partitions = DummyPartitioner(traced_m).propose_partitions()
         partition_nodes = [list(partition.nodes) for partition in partitions]