--- a/torch/fx/passes/utils/fuser_utils.py
+++ b/torch/fx/passes/utils/fuser_utils.py
@@ -249,7 +249,7 @@ def erase_nodes(gm: GraphModule, nodes: NodeList) -> None:
 @compatibility(is_backward_compatible=False)
 def fuse_by_partitions(
     gm: GraphModule,
-    partitions: list[dict[Node, None]],
+    partitions: list[dict[Node, _Optional[None]]],
     prefix: str = "fused_",
     always_return_tuple: bool = False,
 ) -> GraphModule: