Fixes #157891


cc @ezyang @SherlockNoMad @EikanWang @jgong5 @wenzhe-nrv