@xwu-intel 2025-07-09
We hope this could be also cherry-pick to 2.8 so that our customer can use it. Thanks!