@ezyang 2025-07-17
I don't mind maintaining order. Just trying to make sure we're not accidentally quadratic. Perhaps the merging can be done all at once when done?