@xwu-intel 2025-07-18
@ezyang Revised, pls review. The following numbers are from running granite-8b model. 

Perf comparison 
Graph Nodes | Previous fix (seconds) | New fix (seconds) | Speedup 
-- | -- | -- | --
2242 | 0.151868 | 0.038576 | 3.94x
729 | 0.002288 | 0.002368 | 0.97x
2831 | 3.943028 | 3.89862 | 1.01x
1562 | 0.042024 | 0.044988 | 0.93x
2832 | 3.908229 | 4.053367 | 0.96x
1563 | 0.042982 | 0.044441 | 0.97x
2722 | 0.250202 | 0.057822 | 4.33x
732 | 0.001775 | 0.001691 | 1.05x

Overhead of sorting
Graph Nodes | New fix propose_partitions (seconds) | Nodes Sorting Time (seconds) | Sorting Time % of propose_partitionns
-- | -- | -- | --
2242 | 0.038576 | 0.002214 | 5.44%
729 | 0.002368 | 0.000001 | 0.04%
2831 | 3.89862 | 0.000195 | 0.01%
1562 | 0.044988 | 0.000001 | 0.00%
2832 | 4.053367 | 0.000183 | 0.00%
1563 | 0.044441 | 0.000001 | 0.00%
2722 | 0.057822 | 0.006963 | 10.74%
732 | 0.001691 | 0.000000 | 0.00%