@xwu-intel 2025-07-22
@ezyang Lint jobs have script issues or not running?

> ## Merge failed
> **Reason**: 2 mandatory check(s) failed. The first few are:
> 
> * [Lint / lintrunner-noclang / linux-job](https://github.com/pytorch/pytorch/actions/runs/16361773896/job/46410477253)
> * [Lint / lintrunner-mypy / linux-job](https://github.com/pytorch/pytorch/actions/runs/16361773896/job/46410477276)
> 
> Dig deeper by [viewing the failures on hud](https://hud.pytorch.org/pytorch/pytorch/commit/4401ed6db01c54eb1491cffdcf06f51624f4687e)
> 
> Details for Dev Infra team