@xwu-intel 2025-08-04
@ezyang Is the failing caused by this? need fix and retry. 
error: jaxlib 0.6.2 is installed but jaxlib<=0.7.0,>=0.7.0 is required by {'jax'}